save checkpoint after training.
