+++ src/maxdiffusion/configs/base15.yml
@@ -80,8 +80,6 @@ diffusion_scheduler_config: {
   timestep_spacing: ''
 }
 
-# Output directory
-# Create a GCS bucket, e.g. my-maxtext-outputs and set this to "gs://my-maxtext-outputs/"
 base_output_directory: ""
 
 # Parallelism
@@ -158,7 +156,9 @@ max_train_samples: -1
 # max_train_steps takes priority over num_train_epochs.
 max_train_steps: 800
 seed: 0
-output_dir: 'sd-model-finetuned'
+# Output directory
+# Create a GCS bucket, e.g. my-maxdiffusion-outputs and set this to "gs://my-maxdiffusion-outputs/"
+output_dir: ''
 per_device_batch_size: 1
 
 warmup_steps_fraction: 0.0
